% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker.R
\name{isFalsey}
\alias{isFalsey}
\title{Is a value falsey according to Mustache specifications?}
\usage{
isFalsey(x)
}
\arguments{
\item{x}{value}
}
\value{
TRUE if falsey, otherwise FALSE
}
\description{
This function is used to test a value before rendering
}
\keyword{internal}
