\name{rxsplit}
\alias{rxsplit}
\title{Split a character in three parts}
\usage{
rxsplit(x, pattern)
}
\arguments{
  \item{x}{character text to be split}

  \item{pattern}{pattern used for splitting}
}
\description{
It differs from strsplit in that it only splits on the
first occurrence and returns all parts of the string given
}
\keyword{internal}

