% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delim.R
\name{tag2delim}
\alias{tag2delim}
\title{change a delimiter tag into two escaped characters}
\usage{
tag2delim(tag = "{{ }}", escape = TRUE)
}
\arguments{
\item{tag}{character with delimiter tag seperated with a space}
}
\description{
change a delimiter tag into two escaped characters
}
\keyword{internal}
