% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker.R
\name{whisker.escape}
\alias{whisker.escape}
\title{escape basic HTML characters}
\usage{
whisker.escape(x)
}
\arguments{
\item{x}{\code{character} that will be escaped}
}
\value{
HTML escaped character
}
\description{
This method is called for normal mustache keys
}
