% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delim.R
\name{delimit}
\alias{delimit}
\title{enclose a key with delimiters}
\usage{
delimit(x, delim = tag2delim())
}
\arguments{
\item{x}{character with delimiter seperated with a space}

\item{delim}{character vector with escaped delimiters}
}
\description{
enclose a key with delimiters
}
\keyword{internal}
