\name{whisker.render}
\alias{whisker}
\alias{whisker.render}
\title{Logicless templating}
\usage{
  whisker.render(template, data = parent.frame(), debug =
  FALSE)

  whisker(template, data = parent.frame(), debug = FALSE)
}
\arguments{
  \item{template}{\code{character} with template text}

  \item{data}{named \code{list} or env}
}
\value{
  \code{character} with rendered template
}
\description{
  Logicless templating
}

